requirements.txt
Django>=4.2.23
Pillow>=10.0.0
PyPDF2>=3.0.1
pdf2image>=1.16.3
